backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
reportlab>=4.0
gunicorn>=21.2
